Extract SemicolonLogProcessor module
