+++ lib/line-builders.js
@@ -15,12 +15,6 @@ function appendTime (lineParts, { prettified }) {
   }
 }
 
-function appendSemicolon (lineParts) {
-  if (lineParts.length > 0 && lineParts[lineParts.length - 1] !== ':') {
-    lineParts.push(':')
-  }
-}
-
 function appendEOL (lineParts, { EOL }) {
   if (lineParts.length > 0) {
     lineParts.push(EOL)
@@ -54,7 +48,6 @@ function appendObjectOrError (lineParts, { log, EOL, IDENT, errorLikeObjectKeys,
 
 const builtInLineBuilders = {
   appendTime,
-  appendSemicolon,
   appendEOL,
   appendObjectOrError
 }
