+++ lib/log-processors.js
@@ -8,9 +8,9 @@ const IgnoreLogProcessor = require('./processors/IgnoreLogProcessor')
 const LevelLogProcessor = require('./processors/LevelLogProcessor')
 const MessageLogProcessor = require('./processors/MessageLogProcessor')
 const MetadataLogProcessor = require('./processors/MetadataLogProcessor')
+const SemicolonLogProcessor = require('./processors/SemicolonLogProcessor')
 
 const {
-  appendSemicolon,
   appendTime,
   appendEOL,
   appendObjectOrError
@@ -28,12 +28,6 @@ function createLogProcessor (definition) {
     return new builtInLogProcessors[definition]()
   } else {
     return definition
-  }
-}
-
-class SemicolonLogProcessor extends LogProcessor {
-  build (lineParts, context) {
-    return appendSemicolon(lineParts, context)
   }
 }
